Fixed compilation error in 64-bit machines.

+++ c/interface.c
@@ -156,7 +156,7 @@ l_thread(void *param)
 		fflush(stderr);
 	}
 
-	return (void *)ret;
+	return (void *)(long int)ret;
 }
 
 /**
